Code review fix. Removed reference to GitUI assembly.
