+++ Plugins/GitUIPluginInterfaces/UserControls/CredentialsControl.resx
