+++ Plugins/JiraCommitHintPlugin/JiraCommitHintPlugin.csproj
@@ -50,7 +50,6 @@
     <Compile Include="..\..\CommonAssemblyInfo.cs">
       <Link>Properties\CommonAssemblyInfo.cs</Link>
     </Compile>
-    <Compile Include="CredentialsControlBinding.cs" />
     <Compile Include="JiraCommitHintPlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -63,10 +62,6 @@
     <ProjectReference Include="..\..\GitExtUtils\GitExtUtils.csproj">
       <Project>{0F1F1168-A4B2-4FA2-B17B-735140D17F39}</Project>
       <Name>GitExtUtils</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\..\GitUI\GitUI.csproj">
-      <Project>{cf5b22e7-230f-4e50-be88-c4f7023ced2c}</Project>
-      <Name>GitUI</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\ResourceManager\ResourceManager.csproj">
       <Project>{d3440fd7-afc5-4351-8741-6cdbf15ce944}</Project>
